:bug:管理员及员工操作逻辑部分bug修复

+++ app/Model/EmployeeLeave.php
@@ -2,7 +2,7 @@
 /*
  * @Author: xch
  * @Date: 2020-08-15 12:01:16
- * @LastEditTime: 2020-09-14 02:07:45
+ * @LastEditTime: 2020-09-14 12:09:06
  * @LastEditors: Chenhao Xing
  * @Description: 员工信息
  * @FilePath: \epdemoc:\wamp64\www\api-thinkphp\app\Model\Employee.php
@@ -27,13 +27,13 @@ class EmployeeLeave extends Model
     {
         switch ($key) {
             case 'review_satatus':
-                $data = $this->where($key, $value)->where('uuid', $uuid)->paginate($list_rows, $isSimple, $config);
+                $data = $this->where($key, $value)->where('employee_leave.delete_time', null)->where('uuid', $uuid)->paginate($list_rows, $isSimple, $config);
                 break;
             case 'category':
-                $data = $this->where($key, $value)->where('uuid', $uuid)->paginate($list_rows, $isSimple, $config);
+                $data = $this->where($key, $value)->where('employee_leave.delete_time', null)->where('uuid', $uuid)->paginate($list_rows, $isSimple, $config);
                 break;
             default:
-                $data = $this->where('uuid', $uuid)->paginate($list_rows, $isSimple, $config);
+                $data = $this->where('uuid', $uuid)->where('employee_leave.delete_time', null)->paginate($list_rows, $isSimple, $config);
         }
         if (empty($data[0])) {
             return false;
@@ -58,8 +58,8 @@ class EmployeeLeave extends Model
     {
         try {
             $review_status = $this->where('uuid', $uuid)->where('id', $id)->value('review_status');
-            if ($review_status == 2) {
-                return '不能删除已通过审核数据';
+            if ($review_status == 2 || $review_status == 3) {
+                return '不能删除已审核数据';
             }
             //软删除
             $this->destroy($id);
@@ -79,21 +79,21 @@ class EmployeeLeave extends Model
                 $data = Db::view('performance', 'uuid,goods_id,audit_status,create_time')
                     ->view('goods', 'id,goods_name,shop_name', 'goods.goods_id=performance.goods_id')
                     ->where('goods.' . $key, $value)
-                    ->where('employee_leave.delete_time',null)
+                    ->where('employee_leave.delete_time', null)
                     ->paginate($list_rows, $isSimple, $config);
                 break;
             case 'review_status':
                 $data = Db::view('performance', 'uuid,goods_id,audit_status,create_time')
                     ->view('goods', 'id,goods_name,shop_name', 'goods.goods_id=performance.goods_id')
                     ->where('goods.' . $key, $value)
-                    ->where('employee_leave.delete_time',null)
+                    ->where('employee_leave.delete_time', null)
                     ->paginate($list_rows, $isSimple, $config);
                 break;
             default:
                 $data = Db::view('employee_leave', 'id,uuid,category,reason,start_time,end_time,duration,agent,review_status,reviewer,create_time')
                     // ->view('goods', 'order_id', 'goods.goods_id=performance.goods_id', 'LEFT')
                     ->view('employee', 'work_num,real_name', 'employee.uuid=employee_leave.uuid')
-                    ->where('employee_leave.delete_time',null)
+                    ->where('employee_leave.delete_time', null)
                     ->paginate($list_rows, $isSimple, $config);
         }
         if (empty($data)) {
